requirements.txt
annotated-types==0.6.0
anyio==4.2.0
beautifultable==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
distro==1.9.0
docopt==0.6.2
exceptiongroup==1.2.0
filelock==3.13.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
openai==1.12.0
pydantic==2.6.1
pydantic_core==2.16.2
pysondb-v2==2.2.0
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.0
wcwidth==0.2.13
yarg==0.1.9
